pyproject.toml
[project]
name = "coding_farm"
version = "0.1.0"
description = "Antfarm workflows implemented with CrewAI - Multi-agent software development automation"
readme = "README.md"
requires-python = ">=3.10,<=3.13.12"
dependencies = [
    "crewai[litellm,tools]>=0.152.0",
    "click>=8.0.0",
    "gitpython>=3.1.0",
]

[project.scripts]
bug_fix = "bug_fix.main:kickoff"
bug_fix_plot = "bug_fix.main:plot"
feature_dev = "feature_dev.main:kickoff"
feature_dev_plot = "feature_dev.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bug_fix"]

[dependency-groups]
dev = []

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
